References #2, store some of the attrbutes + override displayed user name.
Implemented a hook that would store some of the attirbuted proided by the SSO solution and mostly related to user name data. Language would also be set to the one chosen by user, provided that it is supported. Another hook will replace username data with full name or display name.