References #83, latest Bootstram 4 and tooltip fix.
Upgraded to the latest version of Bootstrap 4, added an override to use the bundle that also includes and internal tooltip.js as Drupal 9 includes version 2, which is not compatible with fourth version of Bootstrap framework. Added a small JS code that would add data-toggle attribute to the sortable table headers.