Resolves #94, implemented a better approach.
Removed previous patch and implemented a better approach that will still give the same result. New approach adds upload max file size configuration to editor, makes sure the option is propagated and adds size in megabytes to the file upload placeholder text. This approach does not rewrite the hub file with a new built and minified version that will not allow to see the changes.